+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
   * [#157](https://github.com/wildland/ember-bootstrap-controls/issues/157): Added new url type input. ([@jweakley][])
   * [#148](https://github.com/wildland/ember-bootstrap-controls/issues/148): Added new text type input. ([@harokevin][])
   * Updated route mixin to better handle adjustable page sizes and associated documentation. ([@jweakley][])
+  * [#153](https://github.com/wildland/ember-bootstrap-controls/issues/153): Added new radio and radio group inputs. ([@jweakley][])
+  * [#149](https://github.com/wildland/ember-bootstrap-controls/issues/149): Added new checkbox input. ([@jweakley][])
 ### Changed
   * Updated `bootstrap-button` to reduce code and take advantage of `drop()` from ember concurrency. ([@jweakley][])
 ### Deprecated